python_code/api/requirements.txt
pandas==2.0.3
python-dotenv==1.0.1
pinecone==5.3.1
openai==1.109.1
runpod==1.7.1